+++ foam/namespace.py
@@ -1,10 +1,11 @@
-__all__ = ['Command', 'Information', 'PostProcess', 'VTK', 'Foam', 'Data', 'Envelope', 'SMTP', 'Figure', 'Timer', 'Version']
+__all__ = ['Command', 'Information', 'PostProcess', 'VTK', 'Foam', 'CaseBase', 'Data', 'Envelope', 'SMTP', 'Figure', 'Timer', 'Version']
 
 
 from .app.command.core import Command
 from .app.information.core import Information
 from .app.postprocess.core import PostProcess, VTK
 from .base.core import Foam
+from .util.object.case import CaseBase
 from .util.object.data import Data
 from .util.object.email import Envelope, SMTP
 from .util.object.figure import Figure