+++ foam/util/function.py
@@ -1,10 +1,15 @@
-__all__ = ['grammar', 'license']
+__all__ = ['dict_without_keys', 'grammar', 'license']
 
 
 import pathlib as p
 
 
 root = p.Path(__file__).parents[1]
+
+
+def dict_without_keys(data: dict, *keys: str) -> dict:
+    # dict := t.Dict[t.Hashable, t.Any]
+    return {key: data[key] for key in data.keys() ^ keys}
 
 
 def grammar() -> str: